requirements.txt
diff-match-patch==20181111
Django==1.10.1
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.1.0
django-shortuuidfield==0.1.3
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.10
Pillow==5.3.0
pytz==2018.7
PyYAML==3.13
shortuuid==0.5.0
six==1.10.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
